:white_check_mark: Update tests for moved log functions
